Improved exit codes. Nicer indentation of status messages.

+++ dld-homepage-mac-com.sh
@@ -4,6 +4,7 @@
 #
 # Usage:   dld-homepage-me-com.sh ${USERNAME}
 #
+# Version 4. Better use of exit codes.
 # Version 3. Now only for homepage.mac.com.
 #    Scrapped PhantomJS, not really necessary for homepage.mac.com.
 # Version 2. Added homepage.mac.com.
@@ -13,16 +14,16 @@
 if [[ ! -x $WGET_WARC ]]
 then
   WGET_WARC=$(which wget)
+  if ! $WGET_WARC --help | grep -q WARC
+  then
+    echo "${WGET_WARC} does not support WARC. Set the WGET_WARC environment variable."
+    exit 3
+  fi
 fi
 
 if [[ ! -x $WGET_WARC ]]
 then
-  echo "wget not found. Set the WGET_WARC environment variable."
-  exit 3
-fi
-if ! $WGET_WARC --help | grep -q WARC
-then
-  echo "${WGET_WARC} does not support WARC. Set the WGET_WARC environment variable."
+  echo "wget-warc not found. Set the WGET_WARC environment variable."
   exit 3
 fi
 
@@ -34,21 +35,21 @@ userdir="data/${username:0:1}/${username:0:2}/${username:0:3}/${username}/homepa
 
 if [[ -f "${userdir}/.incomplete" ]]
 then
-  echo "Deleting incomplete result for ${username}"
+  echo "  Deleting incomplete result for homepage.mac.com/${username}"
   rm -rf "${userdir}"
 fi
 
 if [[ -d "${userdir}" ]]
 then
-  echo "Already downloaded ${username}"
+  echo "  Already downloaded homepage.mac.com/${username}"
   exit 2
 fi
 
 mkdir -p "${userdir}"
 touch "${userdir}/.incomplete"
 
-echo "Downloading ${username}"
-echo -n " - Running wget (takes a while)..."
+echo "  Downloading homepage.mac.com/${username}"
+echo -n "   - Running wget --mirror (takes a while)..."
 
 $WGET_WARC -U "$USER_AGENT" -nv -o "$userdir/wget.log" \
     --directory-prefix="$userdir/files/" \
@@ -57,11 +58,18 @@ $WGET_WARC -U "$USER_AGENT" -nv -o "$userdir/wget.log" \
     --warc-file="$userdir/homepage-mac-com-$username" --warc-max-size=inf \
     --warc-header="operator: Archive Team" \
     --warc-header="mobileme: homepage.mac.com, ${username}"
+if [ $? -ne 0 ]
+then
+  echo " ERROR."
+  exit 1
+fi
 rm -rf "$userdir/files/"
 
 echo " done."
-echo -n " - Result: "
+echo -n "   - Result: "
 du -hs "$userdir/homepage-mac-com-$username"*
 
 rm "${userdir}/.incomplete"
+
+exit 0
